requirements.txt

matplotlib
numpy
pandas
requests
scikit-learn
seaborn
xgboost
statsmodels
seaborn
streamlit
altair
joblib
google.generativeai
pytesseract
tesseract
